v1.3.0: correctness and hardening

+++ src/Tables/Filters/BooleanFilter.php
@@ -18,7 +18,12 @@ class BooleanFilter extends Filter
     public function apply(Builder $query, string $value): void
     {
         if ($this->accepts($value)) {
-            $query->where($this->name, $value === '1');
+            // Bind an int, not a bool. Plenty of schemas store flags as
+            // smallint/integer rather than boolean, and Postgres refuses to
+            // compare those to a bool ("operator does not exist: integer =
+            // boolean"). MySQL and SQLite coerce either way, so binding a bool
+            // worked everywhere the test suite runs and nowhere else.
+            $query->where($this->name, (int) ($value === '1'));
         }
     }
 }
